Added SpeedHack to Cheat Suite
If the player enables the speed hack by pressing F4, a switch statement is used to check the value of the speed/direction memory address.
If the value is 1 (forward) or 255 (backward), the switch statement changes that value to 2 (fast forward) or 254 (fast backward). Left and Right strafe have been commented out - changing their values results in diagonal movement lock.